+++ datafusion/functions-array/src/lib.rs
@@ -31,7 +31,6 @@ pub mod macros;
 mod array_has;
 mod concat;
 mod core;
-mod except;
 mod extract;
 mod kernels;
 mod position;
@@ -57,7 +56,6 @@ pub mod expr_fn {
     pub use super::concat::array_concat;
     pub use super::concat::array_prepend;
     pub use super::core::make_array;
-    pub use super::except::array_except;
     pub use super::extract::array_element;
     pub use super::extract::array_pop_back;
     pub use super::extract::array_pop_front;
@@ -71,6 +69,7 @@ pub mod expr_fn {
     pub use super::replace::array_replace_all;
     pub use super::replace::array_replace_n;
     pub use super::set_ops::array_distinct;
+    pub use super::set_ops::array_except;
     pub use super::set_ops::array_intersect;
     pub use super::set_ops::array_union;
     pub use super::udf::array_dims;
@@ -102,7 +101,6 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         concat::array_append_udf(),
         concat::array_prepend_udf(),
         concat::array_concat_udf(),
-        except::array_except_udf(),
         extract::array_element_udf(),
         extract::array_pop_back_udf(),
         extract::array_pop_front_udf(),
@@ -121,6 +119,7 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         set_ops::array_distinct_udf(),
         set_ops::array_intersect_udf(),
         set_ops::array_union_udf(),
+        set_ops::array_except_udf(),
         position::array_position_udf(),
         position::array_positions_udf(),
         remove::array_remove_udf(),
